feat: break tokenize endpoint out to separate client
